requirements.txt
streamlit==1.32.0
PyPDF2==3.0.1
pdfplumber==0.10.3
requests==2.31.0
python-dotenv==1.0.0
sentence-transformers==2.6.1
faiss-cpu==1.13.2
numpy==1.26.4
groq
